Better description of the write-test-then-write-function workflow that we discussed.
