+++ citrine_converters/old/aramis/converter.py
@@ -6,7 +6,7 @@ from pypif import pif
 import re
 import numpy as np
 import pandas as pd
-from ..tools import replace_if_present_else_append
+from citrine_converters.old.tools import replace_if_present_else_append
 
 
 def ensure_float(x):
