+++ citrine_converters/old/visual/stress_strain.py
