--- forked_code_rbuxman/mechanical/mechanical_properties.py
@@ -1,29 +1,0 @@
-# TODO The pif.System should hold the results--no need for an intermediate
-# object.
-class mechanical_properties():
-    """
-    Class Description:
-    This class will...
-    1. The main purpose of this class is to act as a container to store all the results from calculations.
-    Then at the end this class is called to display the results. I believe a derived function could be
-    implemented to convert the results into whatever format is desired.
-
-    -Apply the Standard here?
-
-needs a panda data frame
-
-Needs to hold:
-Elastic Modulus
-Elastic Onset
-Yield stress
-Yield Strain
-Plastic onset
-Elastic Region
-Plastic Region
-Ultimate Stress
-necking onset
-fracture stress
-total elongation
-ductility
-toughness
-"""
