--- forked_code_rbuxman/test/test_final.py
@@ -1,7 +1,0 @@
-def final_test():
-    """
-    -The purpose of the final tests would be to compare the output of the original code and the forked code
-
-    TBW
-    :return:
-    """    pass
