--- forked_code_rbuxman/test/test_normalizer.py
@@ -1,9 +1,0 @@
-def test_normalizer():
-    """
-    -Check normalizer.py to make sure it correctly normalizes the data
-    -Check for correct unnormalization of data
-
-    TBW
-    :return:
-    """
-    pass
