--- forked_code_rbuxman/test/test_parameters.py
@@ -1,10 +1,0 @@
-def test_parameters():
-    """
-    -Check parameters.py to first of make sure it read and copied everything correct from the config file
-    -Then check and see how good it catches logic errors
-    -Check to see if the assigned defaults don't negatively affect the code
-
-    TBW
-    :return:
-    """
-    pass
