--- forked_code_rbuxman/test/test_units.py
@@ -1,8 +1,0 @@
-def test_units():
-    """
-    -Check units.py to make sure that the data was not changed when going through the function
-
-    TBW
-    :return:
-    """
-    pass
