+++ test/test_hough_transform.py
@@ -1,0 +1,9 @@
+def hough_transform():
+    """
+    -Check hough_transform.py to make sure it gives expected hough and resampled
+    -Check that it yields the correct slope
+
+    TBW
+    :return:
+    """
+    pass
