+++ test/test_stress_strain_curve.py
@@ -1,0 +1,10 @@
+def test_stress_strain_curve():
+    """
+    -Check stress_strain_curve.py to see if it correctly copies and converts the data (into pif and panda)
+    -Check to see if it correctly cuts off the bad end of data
+    -Check to see if it correctly reverses data when in compression
+
+    TBW
+    :return:
+    """
+    pass
